+++ src/ggcmpy/tracing/legacy.py
@@ -198,3 +198,46 @@ class BorisIntegrator_cxx(BorisIntegratorBase):
             fields = df
 
         super().__init__(fields, q, m, integrator_boris_cls=integrator.boris_cxx)
+
+
+class BorisIntegrator_f2py(BorisIntegratorBase):
+    """
+    BorisIntegrator_f2py provides an interface for integrating charged particle trajectories
+    using the Boris algorithm, with field interpolation via f2py-wrapped Fortran routines.
+
+    Args:
+        ds (xr.Dataset or emfields.interpolator_python or emfields.interpolator_yee_python):
+            The dataset containing electromagnetic field data, or a pre-initialized field interpolator.
+        q (float, optional):
+            Particle charge in Coulombs. Defaults to the elementary charge (constants.e).
+        m (float, optional):
+            Particle mass in kilograms. Defaults to the electron mass (constants.m_e).
+
+    Attributes:
+        q (float): Particle charge.
+        m (float): Particle mass.
+
+    Methods:
+        integrate(x0, v0, t_final, dt) -> pd.DataFrame:
+            Integrates the particle trajectory using the Boris algorithm.
+    """
+
+    def __init__(self, df, q=constants.e, m=constants.m_e) -> None:
+        _jrrle.particle_tracing_f2py.boris_init(q, m)
+        if isinstance(df, xr.Dataset):
+            fields = FieldInterpolatorYee_f2py(df)
+        else:
+            assert isinstance(df, FieldInterpolatorYee_f2py)
+            fields = df
+
+        super().__init__(fields, q, m)
+
+    def integrate(self, x0, u0, t_final, dt_max=1.0, dt_max_gyro=0.1) -> pd.DataFrame:
+        n_steps = int(t_final / dt_max) + 2  # add some extra space for round-off issues
+        data = np.zeros((7, n_steps), dtype=np.float32, order="F")
+        n_out = _jrrle.particle_tracing_f2py.boris_integrate(
+            x0, u0, t_final, dt_max, dt_max_gyro, data
+        )
+        return pd.DataFrame(
+            data.T[:n_out], columns=["time", "x", "y", "z", "ux", "uy", "uz"]
+        )
